+++ rqnoj/rqnoj162.cpp
@@ -5,7 +5,7 @@ int main() {
 	scanf("%ld %ld", &n, &m);
 	for(int i=1;i<=m;i++)f[i]=-1000000;
 	for(int i = 0;i < n;i++) {
-		int w,q;
+		long w,q;
 		scanf("%ld %ld", &w, &q);
 		for(int j = w;j <= m;j++) {
 			if(f[j] < f[j-w] + q) {
